feat(main) 글 업데이트 기능 추가, postman json으로 보내는 설정 배움

+++ collect_MrTounge_API/demo/src/main/java/com/mrtounge/demo/service/BoardService.java
@@ -14,5 +14,5 @@ public interface BoardService {
     List<Board> getByListPage(Pageable pageable);
     BoardAndPage getByListAndTotalPage(Pageable pageable);
     Optional<Board> getByListById(Integer id);
-    void updatePost(Long id, String title, String content);
+    void updatePost(Integer id, String title, String content);
 }
